Slightly restructure store methods
