+++ aquadoggo/src/db/mod.rs
@@ -6,7 +6,7 @@ use sqlx::migrate;
 use sqlx::migrate::MigrateDatabase;
 
 pub mod models;
-pub mod sql_storage;
+pub mod store;
 
 /// Re-export of generic connection pool type.
 pub type Pool = AnyPool;
